requirements.txt
# Req2Run Benchmark Dependencies

# Core dependencies
pyyaml>=6.0.1
jsonschema>=4.20.0
pydantic>=2.5.0  # Added to resolve baseline dependency
docker>=7.0.0
kubernetes>=28.1.0
requests>=2.31.0
click>=8.1.7

# Testing and validation
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-timeout>=2.2.0
pytest-cov>=4.1.0

# Performance testing
locust>=2.20.0
aiohttp>=3.9.1

# Security scanning
bandit>=1.7.5
safety>=3.0.1

# Code quality
pylint>=3.0.3
black>=23.12.1
mypy>=1.8.0
flake8>=7.0.0

# Monitoring and metrics
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation>=0.43b0

# Reporting
jinja2>=3.1.2
matplotlib>=3.8.2
pandas>=2.1.4
tabulate>=0.9.0

# Utilities
python-dotenv>=1.0.0
colorlog>=6.8.0
tqdm>=4.66.1
rich>=13.7.0

# Database connectors (for test environments)
psycopg2-binary>=2.9.9
redis>=5.0.1
pymongo>=4.6.1

# Cloud providers SDK (optional)
boto3>=1.34.11  # AWS
google-cloud-storage>=2.13.0  # GCP
azure-storage-blob>=12.19.0  # Azure

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
mkdocstrings>=0.24.0
mkdocstrings-python>=1.7.5